requirements.txt
numpy>=1.24,<3
pandas>=2.0,<3
tqdm
Flask>=2.0
streamlit>=1.30
psycopg2-binary
python-dotenv
